refactor: Move checkIsReadOnlyPermissions in a new permissions helper
